restructure clients
