--- python/flight-fusion/flight_fusion/clients/aio/__init__.py
@@ -1,5 +1,0 @@
-from .inference import AsyncGrpcInferenceServiceClient
-from .repository import AsyncGrpcModelRepositoryServiceClient
-from .utils import run_async
-
-__all__ = ("AsyncGrpcInferenceServiceClient", "AsyncGrpcModelRepositoryServiceClient", "run_async")
