+++ python/flight-fusion/flight_fusion/clients/data/__init__.py
@@ -1,0 +1,6 @@
+from ._dataset import BaseDatasetClient
+from .context import ContextClient
+from .dataset import DatasetClient
+from .versioned import VersionedDatasetClient
+
+__all__ = ("BaseDatasetClient", "ContextClient", "DatasetClient", "VersionedDatasetClient")
