+++ python/flight-fusion/flight_fusion/clients/models/__init__.py
@@ -1,0 +1,3 @@
+from .service import ModelServiceClient
+
+__all__ = ("ModelServiceClient",)
