+++ python/flight-fusion/flight_fusion/clients/models/model.py
